attachable task logic in progress

+++ software/computer/client/qt/expclient/gui/taskstabgui.h
@@ -10,20 +10,20 @@ void tasksTabRefreshTimerUpdateWorker(Ui::MainWindow *ui)
 {
     if (ui->tasksRTMMTableRefreshCheckBox->isChecked())
     {
-        std::map<int, string> rtmms = TaskManager::getInstance()->getRtmms();
+        std::map<long, string> rtmms = TaskManager::getInstance()->getRtmms();
 
         if (rtmms.size() > 0)
         {
             ui->tasksTableWidget->clearContents();
             ui->tasksTableWidget->setRowCount(rtmms.size());
 
-            typedef map<int, string>::iterator it_type;
+            typedef map<long, string>::iterator it_type;
             char i = 0;
             for(it_type iterator = rtmms.begin(); iterator != rtmms.end(); iterator++)
             {
-                int taskId = iterator->first;
+                long taskId = iterator->first;
                 vector<string> taskDesc = Valter::split(((string)iterator->second), '~');
-                ui->tasksTableWidget->setItem(i, 0, new QTableWidgetItem(Valter::format_string("%d", taskId).c_str()));//task id
+                ui->tasksTableWidget->setItem(i, 0, new QTableWidgetItem(Valter::format_string("%lu", taskId).c_str()));//task id
                 ui->tasksTableWidget->setItem(i, 1, new QTableWidgetItem(((string)taskDesc[0]).c_str()));//task name
                 ui->tasksTableWidget->setItem(i, 2, new QTableWidgetItem(((string)taskDesc[1]).c_str()));//task type
                 QTableWidgetItem *item = new QTableWidgetItem(((string)taskDesc[2]).c_str());
